pyproject.toml
[project]
name = "drf-quick-start"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.0",
    "django>=6.0",
    "django-celery-beat>=2.1.0",
    "django-celery-results>=2.6.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "loguru>=0.7.3",
    "mysqlclient>=2.2.7",
    "redis>=7.1.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "djangorestframework-stubs>=3.16.6",
    "drf-spectacular>=0.29.0",
]
